+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [FIX] Objc attributes interop for KMP. See [#1947][]
+
 # 2.14.0 / 04-07-2024
 
 - [IMPROVEMENT] Use `#fileID` over `#filePath` as the default argument in errors. See [#1938][]
@@ -708,6 +710,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#1925]: https://github.com/DataDog/dd-sdk-ios/pull/1925
 [#1934]: https://github.com/DataDog/dd-sdk-ios/pull/1934
 [#1938]: https://github.com/DataDog/dd-sdk-ios/pull/1938
+[#1947]: https://github.com/DataDog/dd-sdk-ios/pull/1947
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu